requirements[1].txt
Flask
pyTelegramBotAPI
aiogram
requests
python-dotenv
